ENH: atleast_nd delegation
